# Conflicts:
#	app/build.gradle
#	app/src/main/java/com/owncloud/android/ui/activity/DrawerActivity.java
#	app/src/main/java/com/owncloud/android/ui/activity/FileDisplayActivity.java
#	app/src/main/java/com/owncloud/android/ui/activity/NotificationsActivity.kt
#	app/src/main/java/com/owncloud/android/ui/activity/SettingsActivity.java
#	app/src/main/res/values-ar/strings.xml
#	app/src/main/res/values-b+en+001/strings.xml
#	app/src/main/res/values-bg-rBG/strings.xml
#	app/src/main/res/values-br/strings.xml
#	app/src/main/res/values-ca/strings.xml
#	app/src/main/res/values-cs-rCZ/strings.xml
#	app/src/main/res/values-da/strings.xml
#	app/src/main/res/values-de/strings.xml
#	app/src/main/res/values-el/strings.xml
#	app/src/main/res/values-es-rAR/strings.xml
#	app/src/main/res/values-es-rEC/strings.xml
#	app/src/main/res/values-es-rMX/strings.xml
#	app/src/main/res/values-es/strings.xml
#	app/src/main/res/values-eu/strings.xml
#	app/src/main/res/values-fa/strings.xml
#	app/src/main/res/values-fi-rFI/strings.xml
#	app/src/main/res/values-fr/strings.xml
#	app/src/main/res/values-ga/strings.xml
#	app/src/main/res/values-gd/strings.xml
#	app/src/main/res/values-hr/strings.xml
#	app/src/main/res/values-hu-rHU/strings.xml
#	app/src/main/res/values-in/strings.xml
#	app/src/main/res/values-is/strings.xml
#	app/src/main/res/values-it/strings.xml
#	app/src/main/res/values-ja-rJP/strings.xml
#	app/src/main/res/values-ka/strings.xml
#	app/src/main/res/values-ko/strings.xml
#	app/src/main/res/values-lo/strings.xml
#	app/src/main/res/values-lt-rLT/strings.xml
#	app/src/main/res/values-mk/strings.xml
#	app/src/main/res/values-nb-rNO/strings.xml
#	app/src/main/res/values-nl/strings.xml
#	app/src/main/res/values-pl/strings.xml
#	app/src/main/res/values-pt-rBR/strings.xml
#	app/src/main/res/values-pt-rPT/strings.xml
#	app/src/main/res/values-ro/strings.xml
#	app/src/main/res/values-ru/strings.xml
#	app/src/main/res/values-sk-rSK/strings.xml
#	app/src/main/res/values-sl/strings.xml
#	app/src/main/res/values-sr/strings.xml
#	app/src/main/res/values-sv/strings.xml
#	app/src/main/res/values-tk/strings.xml
#	app/src/main/res/values-uk/strings.xml
#	app/src/main/res/values-vi/strings.xml
#	app/src/main/res/values-zh-rHK/strings.xml
#	app/src/main/res/values/strings.xml
#	gradle/verification-metadata.xml

+++ app/src/main/java/com/owncloud/android/ui/activities/ActivitiesPresenter.java
@@ -34,7 +34,7 @@ public class ActivitiesPresenter implements ActivitiesContract.ActionListener {
     }
 
     @Override
-    public void loadActivities(int lastGiven) {
+    public void loadActivities(long lastGiven) {
         if (UNDEFINED == lastGiven) {
             activitiesView.showLoadingMessage();
         } else {
@@ -42,7 +42,7 @@ public class ActivitiesPresenter implements ActivitiesContract.ActionListener {
         }
         activitiesRepository.getActivities(lastGiven, new ActivitiesRepository.LoadActivitiesCallback() {
             @Override
-            public void onActivitiesLoaded(List<Object> activities, NextcloudClient client, int lastGiven) {
+            public void onActivitiesLoaded(List<Object> activities, NextcloudClient client, long lastGiven) {
 
                 if (!activityStopped) {
                     activitiesView.setProgressIndicatorState(false);
